Added Ca.crt/key file check condition

+++ ssl.sh
@@ -11,8 +11,12 @@ pass='eltsen'
 HOST_IP=$(ip route get 1 | sed 's/^.*src \([^ ]*\).*$/\1/;q')
 #PUBIP=$(curl https://ifconfig.me/ &> /dev/null)
 
-cacrt=$(curl -SL https://raw.githubusercontent.com/govindinfi/ssl/main/ca.crt -o ca.crt &>/dev/null)
-cakey=$(curl -SL https://raw.githubusercontent.com/govindinfi/ssl/main/ca.key -o ca.key &>/dev/null)
+if [ ! -f "ca.crt" ]; then
+    cacrt=$(curl -SL https://raw.githubusercontent.com/govindinfi/ssl/main/ca.crt -o ca.crt &>/dev/null)
+fi
+if [ ! -f "ca.key" ]; then
+    cakey=$(curl -SL https://raw.githubusercontent.com/govindinfi/ssl/main/ca.key -o ca.key &>/dev/null)
+fi
 
 chmod -R 600 ca.key
 chmod -R 644 ca.crt
